renamed the classname, changed the navigator to component page, rearranged css
