+++ src/app.css
@@ -42,7 +42,7 @@ body {
     height: auto
 }
 
-.wpui_ .wpui_newsletter .nl_container button:hover {
+.wpui_newsletter .nl_container button:hover {
     background-color: #d9d9d9 !important;
     color: #5a6475 !important
 }
@@ -115,7 +115,7 @@ body {
     padding: 40px 20px
 }
 
-div.wpui_preview {
+div.wpui_card_preview {
     background: #b8cfff94;
     border: 10px;
     border-radius: 4px;
@@ -125,24 +125,24 @@ div.wpui_preview {
     justify-content: center;
 }
 
-div.wpui_preview1{
+div.wpui_preview{
    border: 10px solid rgb(255, 255, 255);
    border-radius: 10px;
 }
 
-div.wpui_preview1:hover {
+div.wpui_preview:hover {
     border-color: #e7eefa00;
     background-color:#e7eefa49;
     opacity: 1;
     white-space: nowrap;
 }
 
-div.wpui_preview img {
+div.wpui_card_preview img {
     box-shadow: 0 1px 20px 0 rgba(0, 0, 0, .1), 0 1px 0 -1px rgba(0, 0, 0, .1);
     height: 100%
 }
 
-div.wpui_cl {
+div.wpui_column {
     border-radius: 10px;
     box-shadow: none !important;
     height: auto;
@@ -167,14 +167,14 @@ div.wpui_cl {
 .wpui_view_toggle [role="presentation"] {
     transform: none !important
 }
-.wpui_back_navig span,
-.wpui_back_navig button
+.wpui_back_button span,
+.wpui_back_button button
  {
     color: #535353;
     font-size: .875rem;
     font-weight: 500
 }
-.wpui_back_navig .home:hover{
+.wpui_back_button .home:hover{
     color: #3858e9;
 }
 
@@ -211,7 +211,7 @@ div.wpui_cl {
 }
 
 @media only screen and (max-width:480px) {
-    div.wpui_cl {
+    div.wpui_column {
         column-gap: 10px;
         grid-template-columns: repeat(1, 1fr)
     }
@@ -235,7 +235,7 @@ div.wpui_cl {
 }
 
 @media only screen and (min-width:481px) and (max-width:820px) {
-    div.wpui_cl {
+    div.wpui_column {
         column-gap: 10px;
         grid-template-columns: repeat(2, 1fr)
     }
