+++ src/app.js
@@ -3,8 +3,8 @@
  */
 import { Footer, Header } from './cards';
 import { WpuiProvider } from './data';
-import Navigator from './navigator';
-import ScrollToTop from './scrolltitop';
+import Component_page from './component_page';
+import ScrollToTop from './scrolltotop';
 
 
 /**
@@ -13,12 +13,13 @@ import ScrollToTop from './scrolltitop';
 function App() {
     return (
         <WpuiProvider>
-            <div className='wpui_container'>
-                <ScrollToTop/>
-            <Header/>
-                <Navigator />
-                <Footer />
-            </div>
+            <ScrollToTop>
+                <div className='wpui_container'>
+                    <Header />
+                    <Component_page />
+                    <Footer />
+                </div>
+            </ScrollToTop>
         </WpuiProvider>
     );
 };
